refactor : 승자 출력 메소드 수정
