+++ src/main/java/chess/view/OutputView.java
@@ -60,12 +60,12 @@ public class OutputView {
         System.out.println("> 현재 점수 상태입니다");
         System.out.println("> 화이트 팀 : " + whiteTeamPoint);
         System.out.println("> 블랙 팀 : " + blackTeamPoint);
-        System.out.println(Result.of(team));
+        System.out.println(Result.lose(team));
     }
 
     public static void printKingIsDead(Team team) {
         System.out.println("> 킹이 죽어 경기가 종료되었습니다");
-        System.out.println(Result.of(team));
+        System.out.println(Result.lose(team));
     }
 
     public static void noticeFormerChessBoard(){
